+++ NEWS.md
@@ -34,6 +34,9 @@
 * `format_value()` gains a `decimal_point` argument, to change the decimal point
   in output conversion.
 
+* `format_bf()` with `stars = TRUE` used the `°` symbol for inferiority
+  (evidence *against* the comparison).
+
 * Overhaul of documentation for the package-functions.
 
 ## Bug fix